orders list

+++ frontend/src/app/customer-home/customer-home.component.ts
@@ -2,6 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { ProductsService } from '../products.service';
 import { Product } from '../product';
 import { CartService } from '../cart-service';
+import { Observable } from 'rxjs/Rx';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/observable/throw';
 
 @Component({
   selector: 'app-customer-home',
@@ -25,16 +29,23 @@ export class CustomerHomeComponent implements OnInit {
   }
 
   addToCart(product: Product) {
-    if (!this.cartService.mergeItemToCart(product)){
-      alert("Não é possível adicionar quantidade maior que a disponível em estoque");
-      return;
-    }
-
-    this.totalCart = this.cartService.getTotalValue();
+    this.cartService.mergeItemToCart(product).subscribe(response => {
+      if (!(response === true)){
+        alert("Não é possível adicionar quantidade maior que a disponível em estoque");
+      } else {
+        this.totalCart = this.cartService.getTotalValue();
+      }
+    });
   }
 
-  removeFromCart(product: Product) {    
-    this.cartService.removeItemFromCart(product);
-    this.totalCart = this.cartService.getTotalValue();
+  removeFromCart(product: Product) {
+    this.cartService.removeItemFromCart(product).subscribe(response => {
+      if (response === true){
+        this.cartService.removeItemFromCart(product);
+        this.totalCart = this.cartService.getTotalValue();
+      } else {
+        alert("Não foi possível excluir o item do carrinho");
+      }
+    });
   }
 }
